Cleanup and refactor to use fps

+++ Tilemap2Animation.Test/Services/AnimationEncoderServiceTests.cs
@@ -13,12 +13,9 @@ public class AnimationEncoderServiceTests
     {
         _sut = new AnimationEncoderService();
         _testOutputPath = Path.Combine(Path.GetTempPath(), "test_animation.gif");
-        
+
         // Ensure the test output file doesn't exist before each test
-        if (File.Exists(_testOutputPath))
-        {
-            File.Delete(_testOutputPath);
-        }
+        if (File.Exists(_testOutputPath)) File.Delete(_testOutputPath);
     }
 
     [Fact]
@@ -27,9 +24,9 @@ public class AnimationEncoderServiceTests
         // Arrange
         var frames = new List<Image<Rgba32>>
         {
-            new Image<Rgba32>(16, 16, new Rgba32(255, 0, 0)),
-            new Image<Rgba32>(16, 16, new Rgba32(0, 255, 0)),
-            new Image<Rgba32>(16, 16, new Rgba32(0, 0, 255))
+            new(16, 16, new Rgba32(255, 0, 0)),
+            new(16, 16, new Rgba32(0, 255, 0)),
+            new(16, 16, new Rgba32(0, 0, 255))
         };
         var delays = new List<int> { 100, 100, 100 };
 
@@ -45,26 +42,20 @@ public class AnimationEncoderServiceTests
         finally
         {
             // Clean up
-            foreach (var frame in frames)
-            {
-                frame.Dispose();
-            }
+            foreach (var frame in frames) frame.Dispose();
 
-            if (File.Exists(_testOutputPath))
-            {
-                File.Delete(_testOutputPath);
-            }
+            if (File.Exists(_testOutputPath)) File.Delete(_testOutputPath);
         }
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_WithExistingFile_OverwritesFile()
     {
         // Arrange
         var frames = new List<Image<Rgba32>>
         {
-            new Image<Rgba32>(16, 16, new Rgba32(255, 0, 0)),
-            new Image<Rgba32>(16, 16, new Rgba32(0, 255, 0))
+            new(16, 16, new Rgba32(255, 0, 0)),
+            new(16, 16, new Rgba32(0, 255, 0))
         };
         var delays = new List<int> { 100, 100 };
 
@@ -85,18 +76,12 @@ public class AnimationEncoderServiceTests
         finally
         {
             // Clean up
-            foreach (var frame in frames)
-            {
-                frame.Dispose();
-            }
+            foreach (var frame in frames) frame.Dispose();
 
-            if (File.Exists(_testOutputPath))
-            {
-                File.Delete(_testOutputPath);
-            }
+            if (File.Exists(_testOutputPath)) File.Delete(_testOutputPath);
         }
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_WithEmptyFrames_ThrowsArgumentException()
     {
@@ -105,12 +90,12 @@ public class AnimationEncoderServiceTests
         var delays = new List<int>();
 
         // Act & Assert
-        var exception = await Assert.ThrowsAsync<ArgumentException>(() => 
+        var exception = await Assert.ThrowsAsync<ArgumentException>(() =>
             _sut.SaveAsGifAsync(frames, delays, _testOutputPath));
-        
+
         Assert.Contains("No frames to encode", exception.Message);
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_WithNullFrames_ThrowsArgumentException()
     {
@@ -119,87 +104,78 @@ public class AnimationEncoderServiceTests
         var delays = new List<int> { 100 };
 
         // Act & Assert
-        var exception = await Assert.ThrowsAsync<ArgumentException>(() => 
+        var exception = await Assert.ThrowsAsync<ArgumentException>(() =>
             _sut.SaveAsGifAsync(frames!, delays, _testOutputPath));
-        
+
         Assert.Contains("No frames to encode", exception.Message);
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_WithMismatchedDelaysCount_ThrowsArgumentException()
     {
         // Arrange
         var frames = new List<Image<Rgba32>>
         {
-            new Image<Rgba32>(16, 16, new Rgba32(255, 0, 0)),
-            new Image<Rgba32>(16, 16, new Rgba32(0, 255, 0))
+            new(16, 16, new Rgba32(255, 0, 0)),
+            new(16, 16, new Rgba32(0, 255, 0))
         };
         var delays = new List<int> { 100 }; // Only one delay for two frames
 
         try
         {
             // Act & Assert
-            var exception = await Assert.ThrowsAsync<ArgumentException>(() => 
+            var exception = await Assert.ThrowsAsync<ArgumentException>(() =>
                 _sut.SaveAsGifAsync(frames, delays, _testOutputPath));
-            
+
             Assert.Contains("Delays must match the number of frames", exception.Message);
         }
         finally
         {
             // Clean up
-            foreach (var frame in frames)
-            {
-                frame.Dispose();
-            }
+            foreach (var frame in frames) frame.Dispose();
         }
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_WithNullDelays_ThrowsArgumentException()
     {
         // Arrange
         var frames = new List<Image<Rgba32>>
         {
-            new Image<Rgba32>(16, 16, new Rgba32(255, 0, 0))
+            new(16, 16, new Rgba32(255, 0, 0))
         };
         List<int>? delays = null;
 
         try
         {
             // Act & Assert
-            var exception = await Assert.ThrowsAsync<ArgumentException>(() => 
+            var exception = await Assert.ThrowsAsync<ArgumentException>(() =>
                 _sut.SaveAsGifAsync(frames, delays!, _testOutputPath));
-            
+
             Assert.Contains("Delays must match the number of frames", exception.Message);
         }
         finally
         {
             // Clean up
-            foreach (var frame in frames)
-            {
-                frame.Dispose();
-            }
+            foreach (var frame in frames) frame.Dispose();
         }
     }
-    
+
     [Fact]
     public async Task SaveAsGifAsync_CreatesDirectoryIfNeeded()
     {
         // Arrange
         var frames = new List<Image<Rgba32>>
         {
-            new Image<Rgba32>(16, 16, new Rgba32(255, 0, 0))
+            new(16, 16, new Rgba32(255, 0, 0))
         };
         var delays = new List<int> { 100 };
-        
+
         var nestedPath = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString(), "test_animation.gif");
         var directory = Path.GetDirectoryName(nestedPath)!;
-        
+
         // Make sure directory doesn't exist
-        if (Directory.Exists(directory))
-        {
-            Directory.Delete(directory, true);
-        }
+        if (Directory.Exists(directory)) Directory.Delete(directory, true);
 
         try
         {
@@ -213,15 +189,9 @@ public class AnimationEncoderServiceTests
         finally
         {
             // Clean up
-            foreach (var frame in frames)
-            {
-                frame.Dispose();
-            }
+            foreach (var frame in frames) frame.Dispose();
 
-            if (Directory.Exists(directory))
-            {
-                Directory.Delete(directory, true);
-            }
+            if (Directory.Exists(directory)) Directory.Delete(directory, true);
         }
     }
-} 
+}
